Port navigation to lua
This is the final file. The whole experience feels a lot faster now.
It's not because of the rewrite, but because I've fixed a few dumb
mistakes I made during the initial implementation. It's better now.

There are a few bugs lurking around, but I'll work through them over
time. This is good enough for now.