Add setup function to initialize browser
This handler initializes the plugin and overrides netrw so the consumer
doesn't have to do anything. It doesn't override anything by default so
you can switch back easily without uninstalling. This also gives a space
to add more configuration options in the future, like customizing the
fuzzy finder or mappings.