Add flake file to test changes
This is just an experiment. It allows me to test changes to navitron
without changing my dotfiles build. Instead, it compiles a new editor
with the local changes only in this directory. Now if I mess up the
plugin, it won't break everywhere, just here. I don't have to hack it to
get it working again.