Dynamically link plugin in dev shell
I knew my first attempt at the dev shell wouldn't work long term, but
I wanted to see whether it was at least feasible. The problem was that
local changes would not be reflected until the next commit due to how
Flakes are cached. Now I'm linking to the nearest repo root on startup.
Considering the shell is local to this repo, that seems like
a reasonable bet.