Auto-focus paths you just returned from
If you open a directory and you were just on the `README.md` file, now
your cursor is automatically brought to it. This happens frequently with
the fuzzy finder. This works at any level. If you use the fuzzy finder
and move back up the tree, the selected directory will be the one you
just came from.

This has been bothering me for ages.